refactor: 회원 가입 화면 키, 몸무게 입력 받는 타입 수정

+++ app/src/main/java/com/kjh/dietmanagement/model/Join.kt
@@ -8,9 +8,9 @@ data class Join(
     @SerializedName("password") val password: String,
     @SerializedName("gender") val gender: String,
     @SerializedName("birth") val birth: String,
-    @SerializedName("weight") val weight: Int,
-    @SerializedName("height") val height: Int,
-    @SerializedName("goal_w") val goalW: Int
+    @SerializedName("weight") val weight: Double,
+    @SerializedName("height") val height: Double,
+    @SerializedName("goal_w") val goalW: Double
 )
 
 data class ResponseJoin (
